Apply namespacing and inheritance to cclf claims types
