+++ app/interfaces/api/entities/cclf/final_claim.rb
@@ -1,0 +1,8 @@
+module API
+  module Entities
+    module CCLF
+      class FinalClaim < BaseClaim
+      end
+    end
+  end
+end
