+++ app/interfaces/api/entities/cclf/interim_claim.rb
@@ -1,0 +1,8 @@
+module API
+  module Entities
+    module CCLF
+      class InterimClaim < BaseClaim
+      end
+    end
+  end
+end
