+++ app/interfaces/api/entities/cclf/transfer_claim.rb
@@ -1,0 +1,8 @@
+module API
+  module Entities
+    module CCLF
+      class TransferClaim < BaseClaim
+      end
+    end
+  end
+end
